fix(noodle): give NoodleR a working mobile drawer trigger, trim shell API
NoodleShell now renders its own mobile-drawer opener (a 4th bottom-nav
button) instead of relying on each caller to build one, so NoodlerHome's
previously-unreachable drawer opens on mobile. NoodleShell also derives
hasMorePersonaAccounts, notificationBadgeLabel, and prefersReducedMotion
internally and owns its drawer refs instead of taking them as props,
removing duplicated logic from both callers. Also drops a redundant
viewer-query invalidation in useUpdateNoodlerAccess that the existing
privateAccounts() invalidation already covers via prefix match.

Addresses external review feedback from @Xelvanis on PR #3856.
